+++ src/FinisterreServiceProvider.php
@@ -4,14 +4,12 @@ namespace Buzkall\Finisterre;
 
 use Buzkall\Finisterre\Filament\Livewire\FilterTasks;
 use Buzkall\Finisterre\Filament\Livewire\FinisterreCommentsComponent;
-use Buzkall\Finisterre\Models\FinisterreTag;
 use Buzkall\Finisterre\Models\FinisterreTask;
 use Buzkall\Finisterre\Models\FinisterreTaskComment;
 use Buzkall\Finisterre\Policies\FinisterreTaskCommentPolicy;
 use Buzkall\Finisterre\Policies\FinisterreTaskPolicy;
 use Filament\Support\Assets\Css;
 use Filament\Support\Facades\FilamentAsset;
-use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\Facades\Gate;
 use Livewire\Livewire;
 use Spatie\LaravelPackageTools\Package;
@@ -52,14 +50,6 @@ class FinisterreServiceProvider extends PackageServiceProvider
         Gate::policy(FinisterreTaskComment::class, config('finisterre.comments.model_policy', FinisterreTaskCommentPolicy::class));
 
         $this->loadTranslationsFrom(__DIR__ . '/../resources/lang');
-
-        // Use a custom Tag model to save both Spanish and Catalan translations
-        config(['tags.tag_model' => FinisterreTag::class]);
-
-        // Register morph map so Laravel uses 'tag_id' instead of 'finisterre_tag_id'
-        Relation::morphMap([
-            'tag' => FinisterreTag::class,
-        ]);
 
         // this will get copied to the project's public folder when
         // running php artisan filament:assets